requirements.txt
asgiref==3.2.10
Django==3.1.1
pytz==2020.1
sqlparse==0.3.1
djangorestframework==3.12.2
